Reorganize code, add test
