+++ src/Configuration/ConfigManager.ts
