+++ src/Scene/Ring.test.ts
