+++ src/Scene/Ring.ts
